pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ics-ninja"
version = "1.0.0"
description = "Multi-protocol Industrial Control System security scanner"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "MottaSec", email = "info@mottasec.com"}
]
keywords = ["ics", "scada", "security", "scanner", "ot", "industrial", "modbus", "s7", "opcua", "bacnet", "dnp3"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: System :: Networking",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

# Core dependencies only — protocol-specific libs are optional
dependencies = [
    "click>=8.1.3",
    "colorama>=0.4.6",
    "rich>=13.3.5",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "tabulate>=0.9.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
modbus = ["pymodbus>=3.5.1"]
s7 = ["python-snap7>=1.2"]
mqtt = ["paho-mqtt>=2.1.0"]
opcua = ["opcua>=0.98.13"]
bacnet = ["BAC0>=23.7.3"]
ethernetip = ["pylogix>=0.8.3", "cpppo>=5.2.5"]
profinet = ["scapy>=2.5.0"]
snmp = []  # Uses built-in raw socket implementation
dnp3 = []  # Uses built-in raw socket implementation
hart = []  # Uses built-in raw socket implementation
iec104 = []  # Uses built-in raw socket implementation
network = ["python-nmap>=0.7.1", "scapy>=2.5.0"]
crypto = ["cryptography>=41.0.0"]
all = [
    "pymodbus>=3.5.1",
    "python-snap7>=1.2",
    "paho-mqtt>=2.1.0",
    "opcua>=0.98.13",
    "BAC0>=23.7.3",
    "pylogix>=0.8.3",
    "cpppo>=5.2.5",
    "scapy>=2.5.0",
    "python-nmap>=0.7.1",
    "cryptography>=41.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
ics-ninja = "ics_scanner:cli"

[project.urls]
Homepage = "https://github.com/mottasec/ics-ninja-scanner"
Repository = "https://github.com/mottasec/ics-ninja-scanner"
Issues = "https://github.com/mottasec/ics-ninja-scanner/issues"

[tool.setuptools.packages.find]
include = ["scanners*", "utils*"]

[tool.setuptools]
py-modules = ["ics_scanner"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
